Crate gloo

Gloo is a modular toolkit for building fast and reliable libraries and apps with Rust and WebAssembly.

Modules

console

The JavaScript’s console object provides access to the browser’s console. Using the console object in Rust/WASM directly is cumbersome as it requires JavaScript glue code. This crate exists to solve this problem by providing a set of ergonomic Rust APIs to deal with the browser console.

dialogs

This crate provides wrapper for alert, prompt and confirm functions. web-sys provides a raw API which is hard to use. This crate provides an easy-to-use, idiomatic Rust API for these functions.

events

Using event listeners with web-sys is hard! This crate provides an EventListener type which makes it easy!

file

Working with files and blobs on the Web.

history

A module that provides universal session history and location information.

net

HTTP requests library for WASM apps. It provides idiomatic Rust bindings for the web_sys fetch and WebSocket API.

render

Crate that provides wrapper for requestAnimationFrame

storage

This crate provides wrappers for the Web Storage API

timers

Working with timers on the Web: setTimeout and setInterval.

utils

worker

Workers are a way to offload tasks to web workers. These are run concurrently using web-workers.