Crate gloo::worker

Workers are a way to offload tasks to web workers. These are run concurrently using web-workers.

Communicating with workers

Bridges

After a Worker is spawned, a bridge is created. A Bridge allows bi-directional communication between an worker and a component. Bridges also allow workers to communicate with one another.

Scopes

Scopes are used by workers to communicates with bridges and send updates to itself after a task is finished.

Overhead

Gloo Workers use web workers. They incur a serialization overhead on the messages they send and receive. Bridges use bincode by default to communicate with workers, so the cost is substantially higher than just calling a function.

Structs

Bincode

Default message encoding with bincode.

HandlerId

Identifier to send output to bridges.

WorkerBridge

A connection manager for components interaction with workers.

WorkerDestroyHandle

A handle that closes the worker when it is dropped.

WorkerRegistrar

A Worker Registrar.

WorkerScope

This struct holds a reference to a component and to a global scheduler.

WorkerSpawner

A spawner to create workers.

Traits

Codec

Message Encoding and Decoding Format

Registrable

A trait to enable public workers being registered in a web worker.

Spawnable

A Worker that can be spawned by a spawner.

Worker

Declares the behaviour of a worker.